requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
langchain-community
langchain-openai
python-docx 
pillow
pytesseract
langchain-huggingface
docx